chore: add metrics package
