+++ collector/collector.go
@@ -5,14 +5,6 @@
 package collector
 
 import (
-	"encoding/base64"
-	"encoding/json"
-	"fmt"
-	"io"
-	"log"
-	"net/http"
-	"strings"
-
 	"github.com/dukex/mixpanel"
 	"github.com/google/uuid"
 )
@@ -107,58 +99,4 @@ func (c *collector) CollectError(err error) error {
 	return c.Collect("Error", map[string]interface{}{
 		"details": err.Error(),
 	})
-}
-
-type QueryProfilesResponse struct {
-	Page      int           `json:"page"`
-	PageSize  int           `json:"page_size"`
-	SessionId string        `json:"session_id"`
-	Status    string        `json:"status"`
-	Total     int           `json:"total"`
-	Results   []interface{} `json:"results"`
-}
-
-func getMixpanelUser(distinctId, mixpanelAPISecret, mixpanelPassword string) (*QueryProfilesResponse, error) {
-	url := "https://eu.mixpanel.com/api/2.0/engage?project_id=2729960"
-
-	payload := strings.NewReader(fmt.Sprintf("distinct_id=%v", distinctId))
-
-	req, err := http.NewRequest("POST", url, payload)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("accept", "application/json")
-	req.Header.Add("content-type", "application/x-www-form-urlencoded")
-	req.Header.Add("authorization", fmt.Sprintf("Basic %v", base64.StdEncoding.EncodeToString([]byte(mixpanelAPISecret+":"+mixpanelPassword))))
-
-	res, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	defer res.Body.Close()
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	var mixpanelUser QueryProfilesResponse
-	err = json.Unmarshal(body, &mixpanelUser)
-	if err != nil {
-		return nil, err
-	}
-
-	return &mixpanelUser, nil
-}
-
-func IsFirstTimeUser(distinctId, mixpanelAPISecret, mixpanelPassword string) (bool, error) {
-	user, err := getMixpanelUser(distinctId, mixpanelAPISecret, mixpanelPassword)
-	if err != nil {
-		return false, err
-	}
-
-	log.Printf("USER: %v", user)
-
-	return user.Total == 0, nil
 }
